+++ shared/main/src/commonMain/kotlin/com/jdamcd/runlog/shared/Extensions.kt
@@ -1,6 +1,6 @@
 package com.jdamcd.runlog.shared
 
-expect fun Double.formatKm(): String
+expect fun Float.formatKm(): String
 expect fun Int.formatPace(): String
-expect fun Long.formatDuration(): String
+expect fun Int.formatDuration(): String
 expect fun String.formatDate(pattern: String): String
